(#32) refactor: header,tapbar를 layout으로 이동
